+++ src/samples/mod.rs
@@ -4,4 +4,5 @@
 //! each specification implement the necessary non-generic methods.
 pub mod address_appearance_index;
 pub mod nametags;
+pub mod signatures;
 pub mod traits;