+++ src/specs/mod.rs
@@ -1,6 +1,7 @@
 #![doc = include_str!("../../GETTING_STARTED.md")]
 pub mod address_appearance_index;
 #[allow(dead_code)]
-pub(crate) mod my_data_spec;
+pub(crate) mod my_database;
 pub mod nametags;
+pub mod signatures;
 pub mod traits;