+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `SignaturesSpec` for 4 byte signature support.
+
 ### Changed
 
 - Replaced ssz crate with ssz_rs to have natural constants rather than Typenums.